코드 정리
