+++ client/src/Common/components/ImageSlider.js
@@ -5,7 +5,6 @@ import { Carousel } from 'react-carousel-minimal';
 // import Slider from 'react-slick'
 
 const ImageSlider = (props) => {
-  console.log(props);
   const captionStyle = {
     fontSize: '2em',
     fontWeight: 'bold',
@@ -21,7 +20,6 @@ const ImageSlider = (props) => {
       image: `http://localhost:5000/${image}`
     })
   ))}
-  console.log(data);
   
   return (
     <div>
