+++ client/src/Common/components/SNSImageSlider .js
@@ -2,7 +2,6 @@ import React from "react";
 import { Carousel } from 'react-responsive-carousel';
 
 function ImageSlider(props) {
-  console.log(props);
   return (
     <div>
       <div showArrows={true}>
