+++ client/src/Shop/DetailProductPage/Sections/style/ProductImage.css
@@ -1,0 +1,5 @@
+.videobox {
+  width: 700px;
+  height: 500px;
+  margin: 100px 0 0 300px;
+}
