+++ client/src/Shop/DetailProductPage/style/DetailProductPage.css
@@ -23,5 +23,5 @@
 }
 
 .thumbnail {
-  margin-left: 47px;
+  margin-top: 10px;
 }
