+++ client/src/Shop/LoginMainPage/ShopMainLogin.js
@@ -45,7 +45,7 @@ class MainLogin extends React.Component {
             ref={(ref) => (this.container = ref)}
           >
             <img src={cyworld} style={{ width: "100px" }} />
-            <h1 className="login_page_title">PL2H WORLD</h1>
+            <h1 className="login_page_title">PL2H SHOP</h1>
             {isLogginActive && (
               <LoginPage containerRef={(ref) => (this.current = ref)} />
             )}
