+++ server/mongo/routes/users/cart/userCartInfo.js
@@ -1,7 +1,6 @@
 const express = require("express");
 const router = express.Router();
 const { User } = require("../../../schemas/User");
-const { Cart } = require("../../../schemas/Cart");
 const { Product } = require("../../../schemas/Product");
 
 /////////////////////////////////////////////////////
