+++ server/mongo/routes/users/index.js
@@ -7,10 +7,8 @@ const authRouter = require("./log/auth");
 const addToCart = require("./cart/addToCart");
 const removeFromCart = require("./cart/removeFromCart");
 const userCartInfo = require("./cart/userCartInfo");
-const payment = require("./pay/payment");
-const addCart = require("./sns/addCart");
-const getCart = require("./sns/getCart");
-const getHistory = require("./sns/getHistory");
+const paymentSucces = require("./pay/paymentSucces");
+const paymentgetHistory = require("./pay/getHistory");
 const getMongo = require("./sns/getMongo");
 
 const { auth } = require("../../middleware/auth");
@@ -27,9 +25,6 @@ router.use("/register", register);
 router.use("/login", login);
 router.use("/logout", auth, logout);
 
-router.use("/sns/addCart", addCart);
-router.use("/sns/getCart", getCart);
-router.use("/sns/getHistory", getHistory);
 router.use("/sns/getMongo", getMongo);
 
 /* 장바구니 */
@@ -38,6 +33,7 @@ router.use("/removeFromCart", removeFromCart);
 router.use("/userCartInfo", userCartInfo);
 
 /* 결제 */
-router.use("/payment", payment);
+router.use("/payment/successBuy", paymentSucces);
+router.use("/payment/getHistory", paymentgetHistory);
 
 module.exports = router;
