--- server/mongo/routes/users/sns/getCart.js
@@ -1,17 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const { Cart } = require("../../../schemas/Cart");
-
-/////////////////////////////////////////////////////
-/*                                                 */
-/*          /api/mongo/users/sns/getCart           */
-/*                                                 */
-/////////////////////////////////////////////////////
-
-router.get("/", (req, res) => {
-  Cart.findOne({ email: req.user.email }, (err, cart) => {
-    res.status(200).json({ cart });
-  });
-});
-
-module.exports = router;
